backend/requirements.txt
# URL Shortener Backend Dependencies
# Python 3.12+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database & Cache
asyncpg==0.29.0          # PostgreSQL async driver
redis[hiredis]==5.0.1    # Redis async client with C parser

# Data Validation
pydantic==2.5.3
pydantic[email]==2.5.3

# Environment Variables
python-dotenv==1.0.0

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development Tools (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0            # For testing
black==24.1.1            # Code formatter
flake8==7.0.0            # Linter
